+++ src/mysql_shell/executors/errors/runtime.py
@@ -4,3 +4,8 @@
 
 class ExecutionError(RuntimeError):
     """MySQL shell execution error."""
+
+    # The context attribute is overridden so passwords
+    # are not logged in when showing the traceback
+    # of the previously raised exception
+    __context__ = None