+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog][docs-changelog], and the version adher
 
 
 ## Unreleased
+### Fixed
+- Possible password leak in LocalExecutor traceback info.
 
 ## [0.1.0][changes-0.1.0] - 2025-12-01
 ### Added