+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog][docs-changelog], and the version adher
 - Cluster-set cluster status enum.
 - Complete error message to the LocalExecutor raised exception
 ### Fixed
+- Crash when querying instance role and instance status.
 - Possible password leak in LocalExecutor traceback info.
 
 ## [0.1.0][changes-0.1.0] - 2025-12-01